requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
selenium>=4.10.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
webdriver-manager>=3.8.6